Moved configuration from BootstrapExtension into BootstrapBare
